requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
asgiref==3.8.1
    # via django
asttokens==2.4.1
    # via stack-data
certifi==2024.2.2
    # via sentry-sdk
coverage[toml]==7.5.1
    # via pytest-cov
decorator==5.1.1
    # via ipython
dj-database-url==2.1.0
    # via -r requirements.in
django==5.0.6
    # via
    #   -r requirements.in
    #   dj-database-url
    #   django-debug-toolbar
    #   django-extensions
    #   model-bakery
    #   sentry-sdk
django-debug-toolbar==4.3.0
    # via -r requirements.in
django-extensions==3.2.3
    # via -r requirements.in
django-ordered-model==3.7.4
    # via -r requirements.in
executing==2.0.1
    # via stack-data
flake8==7.0.0
    # via -r requirements.in
iniconfig==2.0.0
    # via pytest
ipython==8.24.0
    # via -r requirements.in
jedi==0.19.1
    # via ipython
matplotlib-inline==0.1.7
    # via ipython
mccabe==0.7.0
    # via flake8
model-bakery==1.18.0
    # via -r requirements.in
packaging==24.0
    # via pytest
parso==0.8.4
    # via jedi
pexpect==4.9.0
    # via ipython
pluggy==1.5.0
    # via pytest
prompt-toolkit==3.0.43
    # via ipython
psycopg[binary]==3.1.19
    # via -r requirements.in
psycopg-binary==3.1.19
    # via psycopg
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pycodestyle==2.11.1
    # via flake8
pyflakes==3.2.0
    # via flake8
pygments==2.18.0
    # via ipython
pytest==8.2.0
    # via
    #   pytest-cov
    #   pytest-django
pytest-cov==5.0.0
    # via -r requirements.in
pytest-django==4.8.0
    # via -r requirements.in
python-decouple==3.8
    # via -r requirements.in
sentry-sdk[django]==2.1.1
    # via -r requirements.in
six==1.16.0
    # via asttokens
sqlparse==0.5.0
    # via
    #   django
    #   django-debug-toolbar
stack-data==0.6.3
    # via ipython
traitlets==5.14.3
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.11.0
    # via
    #   dj-database-url
    #   psycopg
urllib3==2.2.1
    # via sentry-sdk
wcwidth==0.2.13
    # via prompt-toolkit
